Created new BldSendPacket() routine for easier BLD packet support from src code which packs its own BLD packet w/o reading PV's
git-svn-id: file:///afs/slac/g/pcds/vol2/svn/pcds/trunk/pcds/epics/modules/bldClient/current@19415 ececa60f-9f63-0410-97c7-ea05888eb260

+++ bldClientLib/src/bldClientSub.cpp
@@ -42,8 +42,8 @@ static long bldPreSubProcess(subRecord *pSubrecord)
         printf("bldPreSubProcess() : Record %s called bldPreSubProcess(%p)\n",
           pSubrecord->name, (void*) pSubrecord); 
     
-    BldPrepareData(pSubrecord->a);
-        
+    BldPrepareData( static_cast<int>(pSubrecord->a) );
+
     return(0);
 }
